chore(devtools): remove spent audit actuators
Problem
The August 3 devtools-sprawl audit identified four command surfaces whose operator workflows are spent: index fast-forward, archive schema fast-forward, Codex child census, and trajectory reporting. Their catalog entries, wrappers, direct tests, and generated documentation kept retired paths discoverable.

What changed
Remove those four devtools actuators and direct tests, refresh the generated devtools catalog, and extract the reflink clone primitive still required by the live index-v37 actuator into devtools/clone_support.py. Keep the storage index fast-forward lifecycle/executor, their behavior tests, devtools status and verification, beads-state-report, and index-v37-fast-forward.

Compatibility/migration
The four retired devtools commands are no longer discoverable. The beads-state-report graph-health fold and all delete-with-R1 work remain deferred to the later gated wave. Ref polylogue-e2uns.

Verification
Focused managed tests: 41 passed across command catalog, rendered reference, index-v37 actuator, and storage fast-forward lifecycle/executor tests. Real command discovery found 132 commands with the retired commands absent; devtools render devtools-reference --check and devtools status --json passed. The prescribed seed gate passed static/layering/schema/docs checks and reported 18,793 passed, 16 unrelated failures, and 3 skips in its non-slow suite.
